Pull out flatten object to array to its own Utilities file
